refactor: moved str function
