+++ lib/query_processing/src/expressions/functions.rs
@@ -33,6 +33,7 @@ mod starts_ends_contains;
 mod str_;
 mod str_before_or_after;
 mod str_dt;
+mod str_function;
 mod str_lang;
 mod str_len;
 mod struuid;
@@ -41,6 +42,7 @@ mod uuid_v5;
 mod year_;
 
 use crate::errors::QueryProcessingError;
+use crate::expressions::drop_inner_contexts;
 use crate::expressions::functions::abs_::abs_;
 use crate::expressions::functions::ceil_::ceil_;
 use crate::expressions::functions::concat_::concat_;
@@ -75,7 +77,6 @@ use crate::expressions::functions::str_lang::str_lang;
 use crate::expressions::functions::str_len::str_len;
 use crate::expressions::functions::struuid::struuid;
 use crate::expressions::functions::year_::year_;
-use crate::expressions::{cast_lang_string_to_string, drop_inner_contexts};
 use oxrdf::vocab::xsd;
 use oxrdf::NamedNodeRef;
 use polars::datatypes::{DataType, Field};
@@ -84,7 +85,6 @@ use polars::prelude::{
     coalesce, col, lit, Column, Expr, IntoColumn, LiteralValue, Schema, Series, StrptimeOptions,
 };
 use representation::cats::{maybe_decode_expr, LockedCats};
-use representation::multitype::{MULTI_BLANK_DT, MULTI_IRI_DT};
 use representation::query_context::Context;
 use representation::solution_mapping::{BaseCatState, SolutionMappings};
 use representation::{BaseRDFNodeType, RDFNodeState};
@@ -338,59 +338,6 @@ pub fn func_expression(
     }
     solution_mappings = drop_inner_contexts(solution_mappings, &args_contexts.values().collect());
     Ok(solution_mappings)
-}
-
-pub fn str_function(c: &str, t: &RDFNodeState, global_cats: LockedCats) -> Expr {
-    if t.is_multi() {
-        let mut to_coalesce = vec![];
-        for (t, s) in &t.map {
-            to_coalesce.push(match t {
-                BaseRDFNodeType::IRI => maybe_decode_expr(
-                    col(c).struct_().field_by_name(MULTI_IRI_DT),
-                    t,
-                    s,
-                    global_cats.clone(),
-                ),
-                BaseRDFNodeType::BlankNode => maybe_decode_expr(
-                    col(c).struct_().field_by_name(MULTI_BLANK_DT),
-                    t,
-                    s,
-                    global_cats.clone(),
-                ),
-                BaseRDFNodeType::Literal(_) => {
-                    if t.is_lang_string() {
-                        cast_lang_string_to_string(c, t, s, global_cats.clone())
-                    } else {
-                        maybe_decode_expr(
-                            col(c).struct_().field_by_name(&t.field_col_name()),
-                            t,
-                            s,
-                            global_cats.clone(),
-                        )
-                        .cast(DataType::String)
-                    }
-                }
-                BaseRDFNodeType::None => lit(LiteralValue::untyped_null()).cast(DataType::String),
-            })
-        }
-        coalesce(to_coalesce.as_slice()).alias(c)
-    } else {
-        let b = t.get_base_type().unwrap();
-        let s = t.get_base_state().unwrap();
-        match b {
-            BaseRDFNodeType::IRI | BaseRDFNodeType::BlankNode => {
-                maybe_decode_expr(col(c), b, s, global_cats)
-            }
-            BaseRDFNodeType::Literal(_) => {
-                if t.is_lang_string() {
-                    cast_lang_string_to_string(c, b, s, global_cats)
-                } else {
-                    maybe_decode_expr(col(c), b, s, global_cats).cast(DataType::String)
-                }
-            }
-            BaseRDFNodeType::None => lit(LiteralValue::untyped_null()).cast(DataType::String),
-        }
-    }
 }
 
 pub fn xsd_cast_literal(
